update u32 to u64 for variables representating lamports

+++ anchor/programs/solstarter/src/state/contribution.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 #[derive(InitSpace)]
 pub struct Contribution {
-    pub amount: u32,
+    pub amount: u64,
 
     pub user_pubkey: Pubkey,
